frontend/requirements.txt
chainlit==1.3.2
httpx==0.27.2
python-dotenv==1.0.1
